Extract Serializer Attributes into its own file
